+++ gfx/cursors.go
@@ -119,5 +119,3 @@ func (c *Cursors) ConvertMouseClickToTextCursorPosition(button, action uint8) {
 		}
 	}
 }
-
-
