--- main.go
@@ -1,107 +1,0 @@
-/*
-
-------- TODO: -------
-
-* KEY-BASED NAVIGATION
-	* CTRL-HOME/END - PGUP/DN
-* BACKSPACE/DELETE at the ends of lines
-	pulls us up to prev line, or pulls up next line
-* when there is no scrollbar, should be able to see/interact with text in that area
-* when auto appending to the end of a text panel, scroll all the way down
-		(manual activity in the middle could increase size, so do this only when appending to body)
-
-
-------- LOWER PRIORITY POLISH -------
-
-* if typing goes past right of screen, auto-horizontal-scroll as you type
-* same for when newlines/enters/returns push cursor past the bottom of visible space
-* scrollbars should have a bottom right corner, and a thickness sized background
-		for void space, reserved for only that, so the bar never covers up the rightmost
-		character/cursor
-* when pressing delete at/after the end of a line, should pull up the line below
-* vertical scrollbars could have a smaller rendering of the first ~40 chars?
-		however not if we map the whole vertical space (when scrollspace is taller than screen),
-		because this requires scaling the text.  and keeping the aspect ratio means ~40 (max)
-		would alter the width of the scrollbar
-
-*/
-
-package main
-
-import (
-	"fmt"
-	"go/build"
-	_ "image/png"
-	"log"
-	_ "os"
-	"runtime"
-
-	"github.com/corpusc/viscript/app"
-	"github.com/corpusc/viscript/gfx"
-	"github.com/corpusc/viscript/hypervisor"
-	"github.com/corpusc/viscript/msg"
-	"github.com/go-gl/gl/v2.1/gl"
-	"github.com/go-gl/glfw/v3.2/glfw"
-)
-
-func init() {
-	fmt.Println("main.init()")
-	gfx.MakeHighlyVisibleLogHeader(app.Name, 15)
-	// GLFW event handling must run on the main OS thread
-	// See documentation for functions that are only allowed to be called from the main thread.
-	runtime.LockOSThread()
-}
-
-func main() {
-
-	fmt.Printf("Start\n")
-
-	if err := glfw.Init(); err != nil {
-		log.Fatalln("failed to initialize glfw:", err)
-	}
-
-	defer glfw.Terminate()
-
-	glfw.WindowHint(glfw.Resizable, glfw.True)
-	glfw.WindowHint(glfw.ContextVersionMajor, 2)
-	glfw.WindowHint(glfw.ContextVersionMinor, 1)
-	window, err := glfw.CreateWindow(int(gfx.CurrAppWidth), int(gfx.CurrAppHeight), app.Name, nil, nil)
-
-	if err != nil {
-		panic(err)
-	}
-
-	window.MakeContextCurrent()
-
-	if err := gl.Init(); err != nil {
-		panic(err)
-	}
-	hypervisor.Texture = hypervisor.NewTexture("Bisasam_24x24_Shadowed.png")
-
-	defer gl.DeleteTextures(1, &hypervisor.Texture)
-
-	hypervisor.InitRenderer()
-
-	hypervisor.InitInputEvents(window)
-
-	for !window.ShouldClose() {
-		msg.MonitorEvents(hypervisor.Events)
-		glfw.PollEvents()
-		hypervisor.DrawScene()
-		//drawScene()
-		window.SwapBuffers()
-	}
-}
-
-// importPathToDir resolves the absolute path from importPath.
-// There doesn't need to be a valid Go package inside that import path,
-// but the directory must exist.
-func importPathToDir(importPath string) (string, error) {
-	p, err := build.Import(importPath, "", build.FindOnly)
-
-	if err != nil {
-		return "", err
-	}
-
-	return p.Dir, nil
-}
